Various tweaks for planes
Fixed the planes' collider so they don't immediately depart, adjusted some stuff in their behavior so they land appropriately and don't try to keep flying after landing. Also adjusted the taxi and launch paths in the hangar so the planes don't stick their noses out the back. Something seems weird with those paths in the prefab versus the scene but moving them in the scene seems to be the right thing to do?