Made some tweaks to the launch UI system
Adjusted how the available task list was being created to align more with how tasks are managed by the GameManager. Also edited the Runway Prefab to make it so the Hanager is the only LandingLocation so you can't click on the individual runways anymore. It seems to work though I think the 0 point of the Runway prefab is off.